miniproject2/profolder/project2/requirements.txt
ultralytics>=8.0.0
opencv-python
google-generativeai
pyttsx3
colorama
numpy
pillow
flask
python-dotenv
